+++ neopdf/src/lib.rs
@@ -2,10 +2,12 @@
 
 pub mod converter;
 pub mod gridpdf;
-pub mod interpolation;
+pub mod interpolator;
 pub mod manage;
 pub mod metadata;
 pub mod parser;
 pub mod pdf;
+pub mod strategy;
+pub mod subgrid;
 pub mod utils;
 pub mod writer;
