+++ neopdf/src/strategy.rs
