+++ neopdf_pyapi/src/gridpdf.rs
@@ -1,7 +1,9 @@
 use ndarray::Array1;
-use neopdf::gridpdf::{GridArray, ParamRange, SubGrid};
 use numpy::{PyArrayMethods, PyReadonlyArray6};
 use pyo3::prelude::*;
+
+use neopdf::gridpdf::{GridArray, ParamRange};
+use neopdf::subgrid::SubGrid;
 
 /// Python wrapper for the `SubGrid` struct.
 #[pyclass(name = "SubGrid")]
